internet-latency: use shared partitioned buffer
